+++ vllm/triton_utils/__init__.py
@@ -17,4 +17,7 @@ else:
     tl = TritonLanguagePlaceholder()
     tldevice = TritonLanguagePlaceholder()
 
-__all__ = ["HAS_TRITON", "triton", "tl", "tldevice"]
+LOG2E = tl.constexpr(1.4426950408889634)
+LOGE2 = tl.constexpr(0.6931471805599453)
+
+__all__ = ["HAS_TRITON", "triton", "tl", "tldevice", "LOG2E", "LOGE2"]
